+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## [Unreleased]
 
  - 📝 Add model in the exception message when an Image does not have a defined size
+ - 🐛 `ButtonContrast` and `ButtonGhost` were using `color` instead of `onContainerColor` values for their content colours
+ - 📝 Improve the message from the crash when using a Spark component outside a `SparkTheme`
 
 ## [1.8.0-alpha03]